// graphql-src/blog-api.ts
export type Blog = {
  __typename: "Blog",
  id: string,
  title: string,
  summary?: string | null,
  content?: string | null,
  timestamp?: string | null,
  likes?: number | null,
  createdAt: string,
  updatedAt: string,
};